renamed folder under services to aws
